Print out a meaningful error when the state file can't be opened
git-svn-id: https://labs.omniti.com/resmon/branches/resmon2@379 8c0face9-b7db-6ec6-c4b3-d5f7145c7d55

+++ lib/Resmon/Status.pm
@@ -15,7 +15,10 @@ my $REQUEST_TIMEOUT = 60;
 sub new {
     my $class = shift;
     my $file = shift;
+    # State file used for communication between monitor and webserver
+    # processes
     my $fh = IO::File->new(".$file.state", "+>");
+    die "$0: Unable to open .$file.state: $!\n" unless (defined $fh);
     # Delete the just opened file - it stays open, but doesn't show on disk
     unlink ".$file.state";
     return bless {
